requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic[email]
pydantic-settings
python-jose[cryptography]
bcrypt
python-multipart
boto3
celery[redis]
redis
httpx

